+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - (Feature) (DebugPackage) CrashHandler Data
 - (Feature) (DebugPackage) Activities
 - (Feature) (Platform) Package Copy CLI
+- (Feature) (Platform) MetaV1 GetBatch
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService